+++ editors/jetbrains/src/main/kotlin/io/agnix/jetbrains/lsp/AgnixLspServerSupportProvider.kt
@@ -3,6 +3,7 @@ package io.agnix.jetbrains.lsp
 import com.intellij.openapi.project.Project
 import com.redhat.devtools.lsp4ij.LanguageServerFactory
 import com.redhat.devtools.lsp4ij.client.LanguageClientImpl
+import com.redhat.devtools.lsp4ij.installation.ServerInstaller
 import com.redhat.devtools.lsp4ij.server.StreamConnectionProvider
 
 /**
@@ -19,6 +20,10 @@ class AgnixLspServerSupportProvider : LanguageServerFactory {
 
     override fun createLanguageClient(project: Project): LanguageClientImpl {
         return AgnixLanguageClient(project)
+    }
+
+    override fun createServerInstaller(): ServerInstaller {
+        return AgnixLspServerInstaller()
     }
 
     /**